Support pinning a kernel
The ZFS filesystem option 'org.zfsbootmenu:kernel' can now be set on a
boot environment. If set, this pattern is used to match a kernel file
name in /boot .

zfs set org.zfsbootmenu:kernel=5.3.10 zroot/ROOT/void.2019.11.15

If zroot/ROOT/void.2019.11.15 is your automatic boot environment (bootfs
on the pool), the first-found kernel in /boot that matches '5.3.10' will
be booted.  In this example, 'vmlinuz-5.3.10_1' will match and be
booted.

This match only applies in the automatic boot menu. If you escape to the
environment/kernel/snapshot selection menus, the most recent kernel (or
one you specifically pick in a menu) will be used.