Split functions to include, document, DRY
- Move all functions to a file sourced in zfsbootmenu.sh.
- Document the arguments, return code and output of each function.
- Exercise DRY, now that the overal structure is more stable - new
functions were created to condense duplicated logic, existing functions
were modified to do a bit more work.

- Snapshot browsing now clones the snapshot, and brings you back to the
BE selection menu, so you can boot the snapshot or select a specific
kernel.

- Automatic booting and booting via the boot menu now prefer a kernel
version if org.zfsbootmenu:kernel is set on the filesystem. This
behavior will result in fewer surprises when booting.