Fix edge cases
If zpool is set to a non-existent pool on the kernel command line, we
incorrectly handle setting a kernel mask in select_kernel.

This also simplifies setting BOOTFS, so we should now always have it
correctly set, or empty.

Add the 'ip' tool, so you can set an IP and zfs send/recv in the
bootloader.