Trap kernel module installation failure
Instead of silently ignoring installation failures for critical kernel
modules, we now check the return code from the install helper for key
modules. If the module could not be installed, the error is trapped and
an exit is forced to propagate this up to generate-zbm, which in turn
exits before replacing or removing any initramfs/efi files.

Optional modules can still have installation attempts without erroring
out the process.