Initial support for org.zfsbootmenu:active visibility
If a BE has mountpoint=/, setting org.zfsbootmenu:active=off will hide
the BE from the boot menu; all other values will cause it to be shown.

Conversely, if a BE has mountpoint=legacy, setting
org.zfsbootmenu:active=on will cause the BE to be shown in the boot
menu; all other values will prevent its display.