Allow root prefix to be customized for other distributions
Fixes #21.

When creating a `root=` kernel command-line, kexec_kernel now defers to
a function that will attempt to determine the appropriate prefix to
prepend to the ZFS dataset name. If the `org.zfsbootmenu:rootprefix`
property is defined on the BE, that will be preferred; otherwise, if
`/usr/lib/os-release` or `/etc/os-release` are available, those will be
queried for OS identifiers will be searched for a preexisting default.
If no suitable values are discovered, a default of "zfs:" is used.