Add string interpolation to ZBM version handling
The '--version' command-line argument now gets a Global.Version YAML
configuration option to allow the ZBM version to be set by default. Both
Global.Version (which governs output versions) and Kernel.Version (which
governs input versions) support limited string interpolation: the value
"%current" (when ending on a word boundary) or "%{current}" (without
regard to a final word boundary) will be replaced. In Global.Version,
"%{config}" becomes the value of the $VERSION variable in generate-zbm.
In Kernel.Version, "%{config}" becomes the value of `uname -r`.

The Global.Version behavior was already established in a more limited
form as replacing `^%current$` with `uname -r`.