Add documentation for generate-zbm, its config and initramfs options
Master documents reside in the `pod` subdirectory (except for
`generate-zbm.8`, which is inlined in `bin/generate-zbm` and may be
converted to man pages by running, from the repository root, the helper
`releng/pod2man.sh`. The `releng/tag-release.sh` helper will convert the
docs as part of the release process.

Co-authored-by: Zach Dykstra <dykstra.zachary@gmail.com>
Co-authored-by: Andrew J. Hesford <ajh@sideband.org>
Co-authored-by: Érico Rolim <erico.erc@gmail.com>