Drop import_args variable and let import_pool build its own arguments
Rather than define `import_args` in zfsbootmenu-parse-commandline.sh
based on the value of `force_import` and `read_write` on the ZBM kernel
command line, the parser just defines `force_import` and `read_write`
variables based on these arguments. The `import_pool` function uses
these variables directly to determine the right command-line arguments
to pass to `zpool import`.

This is a bit cleaner and allows for a construct like

    read_write=yes import_pool zpool

to force rw imports rathern than rewriting `import_args`. Note that
passing `read_write=yes` in that call does not override the value in the
environment, so other calls to `import_pool` will fall back to expected
behavior.

Closes #75.