Support automatic "noresume" when importing pools R/W
Rather than just warn the user about the need to set the "noresume"
kernel argument when booting after an active suspend partition is
detected and a pool is imported R/W, ZFSBootMenu now supports rewriting
the kernel command line to remove any "resume=" arguments and append a
"noresume". This uses awk to cleanly (and, it seems, correctly) handle
the possibility of double-quoted strings in the argument list, which the
kernel docs say are permitted. (The docs do not mention single-quoted
strings, so we don't support them.)

The prior behavior has been preserved by allowing the user to type
DANGEROUS instead of the redefined NORESUME response to the resume guard
prompt, just in case there is some unexpected edge case that thwarts
automatic redirection and the user wants to assume responsibility for
pool corruption.

Closes #78.