Allow ctrl-c to be used during user input
Globally disable ctrl-c in /bin/zfsbootmenu by pointing it to an empty
handler. Inside a subshell, unset the empty handler and then invoke
zfsbootmenu-input with a single quoted argument; the pre-filled line
that the user can edit. If ctrl-c is detected here, the script kills
itself. Otherwise, it echos the line the user entered.

It is up to the caller to determine how to handle an empty response.