testing: limit depth in dracut module search, drop `--local` option
It turns out that `dracut --local` doesn't just alter the location of
the `modules.d` directory, it expects to find (and only works if it
finds) *all* of the dracut components in a local tree. As it was used in
the test setup, it was effectively a no-op.

Because we really don't want to have to pull in all of dracut, dropping
the `--local` option makes the most sense.