Rewrite pool import ahead of OpenZFS 2.0.0
The historical pool import process relied on scraping the pool status
and attempting to import any pool that is marked as ONLINE. This process
fails for pools with feature flags enabled from a newer version of ZFS
than what is in the initramfs. The process has been reworked as follows:

* Rely on zpool to import every pool possible in read-only mode
* If any pool has unavailable feature flags, set a global flag
  disabling read-write pool operations for that pool, and
  then provide a warning indicating that ZFS in the initramfs
  should be upgraded.
* If the preferred zpool could not be imported, warn and then
  drop to a recovery shell.

Additionally, the read_write kernel command line option has been
removed, as this behavior conflicts with the requirement to import
pools read-only initially.

Closes #87

Co-authored-by: Andrew J. Hesford <ajh@sideband.org>
Co-authored-by: Zach Dykstra <dykstra.zachary@gmail.com>