Embedded help system
To enable a global help system, bind alt-h by default to all instances
of fzf and sk. The bind executes zfsbootmenu-help -L $_SECTION.

Before invoking each instance of fzf/sk, set _SECTION to the appropriate
value (MAIN, SNAPSHOT, ...). zfsbootmenu-help then uses this argument to
pre-sort the list of topics, pushing that topic to the head of the list
/ pre-selecting it.

The documentation is internal to zfsbootmenu-help, and is created using
a very simple templating system with basic colors and text centering.
fold is used to word wrap based on the size of the preview window. The
preview window size is calculated based on the number of columns minus
26 characters (the longest topic title + fzf marker ).

When adding new help sections, care should be taken with colorizing
text. Because ANSI escape sequences are used, fold is not able to
correctly calculate the length of given word, so it word wraps
incorrectly.