Enable dynamic line wrapping for header
Terminal/console widths vary in size depending on the hardware in a
system. Because of this, we need to automatically wrap the shortcut key
header (positioned as a footer) when it overflows the console width.

header_wrap() accepts 1 or more arguments. Each argument is considered a
token that can't be split/wrapped.

To increase visual contrast, [WORD] is colored green, with the remainder
of the token colored light blue.

Closes #93

Co-authored-by: Zach Dykstra <dykstra.zachary@gmail.com>
Co-authored-by: Andrew J. Hesford <ajh@sideband.org>