Sort environments, fix preview
Fix a string comparison in zfsbootmenu-preview.sh to make it an exact
match instead of a regexp match. This will prevent multiple environments
showing up as (default).

In populate_be_list, reverse sort the list of candidate boot
environments and then append the BOOTFS value if defined. Combined with
fzf/sk --tac, this pushes the boot environment to the top of the list so
that it is the first selected item on the main screen.

Closes #95