Fix console handling when attached to a serial line
When attaching the console to a serial line (as seen in the testing
setup), booting a system from ZFSBootMenu results in an invisible cursor
on the serial console. Invoking `tput cnorm` to restore normal cursor
visibility in `kexec_kernel` restores the cursor on the serial line.

Setting `control_term=/dev/tty0` in zfsbootmenu-parse-commandline.sh
seems to cause issues with job control in emergency_shell when a serial
console is also attached. Using `control_term=/dev/tty1` instead appears
to fix this problem. Accordingly, `console=tty1` is also preferred in
the kernel command line of the testing setup.