Loop the emergency shell when initial pool imports fail
Originally, ZFSBootMenu would drop to an emergency shell once if the
initial `zpool import -aN` failed to yield a usable pool, then terminate
when the user quit the shell. This required the user attempt to make the
pool importable and then reboot to allow ZFSBootMenu to try again.

With this change, the pool health check and drop to emergency_shell are
looped until a usable pool is identified. This allows the user to
manually import a pool if possible, then type `exit` to allow
ZFSBootMenu to attempt to proceed without an intervening reboot.