Fall back to legacy hostid creation for ZFS < 2.0
The dracut module in ZFS < 2.0 produces initramfs images with all-zero
hostid files on musl systems, regardless of the contents of /etc/hostid
on the host. In ZFS >= 2.0, the dracut module respects the contents of
/etc/hostid when possible, and will otherwise create a new /etc/hostid
only if the hostid(1) program returns a non-zero value.

To minimize the possibility of import failures, ZFSBootMenu should try
hard to replicate the behavior of whichever version of ZFS is creating
initramfs images.

The version comparison here should be considered deprecated and removed
once ZFS >= 2.0 has become sufficiently widespread.