Initial split of core menu logic from initialization
To support remote interaction with something like dropbear, the core
menu should be divorced from the early setup and executable from the
command line. The setup is done exactly once and touches an
"initialized" file to signal when it has finished. The menu script
awaits this signal when it launches. To prevent simultaneous execution
of multiple instances, the menu script uses a mutex.