Improve key handling
The key_wrapper function has been collapsed into load_key, which tries
to do the right thing when handling encryptionroot and unencrypted
filesystems. The mount_zfs function now also checks if a filesystem is
locked and, if so, fails the mount; because mount_zfs tends to be run in
a subshell, attempting to mount a locked dataset will hang.