Optionally launch under tmux
To facilitate live-initramfs debugging, include the dracut-tmux module
and set zbm.tmux=1 on the ZFSBootMenu kernel command line. This will
launch zfsbootmenu-countdown.sh / zfsbootmenu.sh under tmux, enabling
you to easily launch another shell in the initramfs. This is a
double-opt in, so that tmux isn't randomly installed in an initramfs
unless explicitly desired.