Support optional "teardown" script to run before kexec
A new `zfsbootmenu_teardown` dracut option has been added to
`module-init.sh` that, when set to a path that refers to an executable,
will install the named executable as `/libexec/zfsbootmenu-teardown` in
the initramfs. In `kexec_kernel`, if `/libexec/zfsbootmenu-teardown` is
found to exist and be executable, the script will be run immediately
before the final `kexec` call to jump into the new kernel.