Support configurable, opt-in caching of key files
By setting org.zfsbootmenu:keysource=<keysource> on a BE or its parent,
load_key will attempt to look for a key file specified by a file://
keylocation relative to the filesystem specified by <keysource> and will
cache this key within the initramfs. This allows single-sign-on for
multiple BEs that have different encryption roots but a common keysource
and keylocation.