Change chroot keys, add help text
On the main screen and the snapshot screen, normalize the chroot key
combination to MOD-i (interactive chroot). Help documentation has been
updated for the main screen and added for the snapshot screen.

Additionally, document MOD-w and list the caveats for when it can be
used. This key is not documented in the header text - read the
documentation to understand what kind of foot-gun you're loading.