Normalize tests for paths.
The `-f` flag to test(1) checks for regular files, which is generally
not the right thing to do when looking for ZFS keys because keylocation
can point to, e.g., a device node.

Every `-f` test in where we expressly intend to read (or source) a file
has been moved to `-r`, and tests for ZFS keys (which are intended to be
passed as locations to `zfs load-key` rather than read directly) are
relaxed to the maximally generic existence test (`-e`).

Also, swap `test` for `[ ]` syntax everywhere and drop sourcing of
`zfsbootmenu-lib.sh` in the PWD. That was probably an artifact of early
testing, but we have better test setups now. Instead, look only for the
canonical `/lib/zfsbootmenu-lib.sh`.

Closes #127.