Expect names of the form <prefix>-<version> when finding latest kernel
If a specific kernel version is not defined in config.yaml, the name of
the kernel is split on the first '-' to determine the prefix and the
version. If no version can be found, this is a hard error.

The latestKernel subroutine is only called if neither a kernel path nor
the kernel version is specified, which means its output will always be
split on '-' to identify a prefix and a version. Thus, in the
subroutine, only look for files of the form <prefix>-*, silently
ignoring kernels with non-conforming names. This allows for clean
operation if at least one kernel does match the expected pattern, and
also provides for more meaningful error messages if no kernels match the
expected pattern (for example, on Arch).

Fixes #128.