Improve kernel version detection.
The dracut lsinitrd utility determines kernel versions by looking for
strings of the form `[0-9]+\.[0-9]+\.[0-9]+`, so we might as well adopt
similar logic to try to recover versions when the file name is not
sufficient. This improvement attempts to parse both the file name and
kernel strings for version-like sequences. If it finds both, they must
agree or the kernel is ignored; if it finds only one source of
versioning, that is preferred automatically.

With more reliable version detection, we can again relax the kernel
detection to look for `/boot/<prefix>` rather than requiring the more
strict `/boot/<prefix>-<version>`, and an output kernel prefix becomes
just the name of the kernel with the determined version removed (if it
exists).

This *should* allow Arch users to rely on automatic version detection
rather than requiring explicit specification of the kernel path and
version. This also allows Arch users to run `generate-zbm` after
updating a kernel package but before rebooting, when using a
`%{current}` version string will fail because `/lib/modules/$(uname -r)`
has been removed and the kernel file is of a different version anyway.

Closes: #132.