Only specify `--uefi-stub` when EFI.Stub is configured
When `--uefi` is passed to dracut without a `--uefi-stub` argument, it
looks first for a stub provided by systemd-boot and, if one isn't found,
for a stub provided by gummiboot. The gummiboot location agrees with our
default location at /usr/lib/gummiboot/linuxx64.efi.stub.

By dropping the `--uefi-stub` unless the `generate-zbm` configuration
explicitly specifies a value, `generate-zbm` should work out of the box
on systems that use gummiboot as well as its successor, systemd-boot. If
a user specifies a value in EFI.Stub, the `--uefi-stub` argument will be
passed with the configured value.

Fixes: #135.
Closes: #138.