Allow BE and snapshot sorting by different criteria  (#143)
On systems with a large number of boot environments or snapshots, sorting by name can be less than optimal. This PR adds the following:

    Allow zbm.sort_key to be set on the KCL, with name, creation, used considered valid sort keys
    Add a hook for mod-o on the BE and snapshot list screens, allowing the sorting method to be changed. Changing the sorting method on one screen changes it on the other, as well. This change persists until zfsbootmenu.sh is exited, at which point the selected method reverts to the default (name) or the method defined on the KCL.

Packing the sort methods into the global zbm_sort variable allows easy expansion of sort keys in the future. Simply add the list of valid keys to the array in parse-commandline.sh and it will be accepted and rotated through the list.

The shortcut key combinations have been documented in the online help and the man page, but not in the footer text.