Support old versions of craaaaap
Check if dmesg supports the '--noescape' flag
Check if fzf supports the 'refresh-preview' action

Create /etc/zfsbootmenu.conf in the initramfs with exports, so the file
can be sourced in the initramfs. Variables in here should be considered
feature flags - if the variable is defined with any value, the flag is
set to true. If the variable is undefined, consider the feature
disabled.

BYTE_ORDER: influence how /etc/hostid is written
HAS_NOESCAPE: enable color escape codes in debug-level messages, parse
them in zlogtail
HAS_REFRESH: fzf is new enough to support the refresh-preview action.
This has the added benefit of un-breaking skim support, again