Common network config and ZBM setup in non-Void VMs
Because all non-Void VMs currently use systemd, they now use a common
setup script to configure systemd-network and systemd-resolved for
network access. In addition, all non-Void VMs install ZBM prerequisites
and pre-populate a clone of the repo at /zfsbootmenu. The ZBM install
taret is run, but the installed 90zfsbootmenu and generate-zbm
components are replaced with symlinks to the repo copy to facilitate
testing builds from other branches.

The ZBM config is preset to manage component images without versioning
and store the outputs in /zfsbootmenu/build for easy extraction to the
host.