Remove dracut-ism, replace with an internal check
/lib/dracut/hooks/initqueue/finished is rm -rf'd when rd.hostonly=0 is
set, resulting in an on screen error when the echo is executed. If for
some reason /dev/zfs is not present, under the old setup, a dracut
initqueue timeout would happen and you'd drop down to a limited dracut
shell. A check for /dev/zfs is performed just before any zpool work is
attempted and our emergency shell is instead used.