Store options from KCL in a file for easy sourcing
zfsbootmenu-parse-commandline.sh defines several configuration variables
based on KCL arguments; zfsbootmenu-exec.sh originally exported those to
the environment for use by zfsbootmenu-countdown.sh and zfsbootmenu.sh.
This works fine if zfsbootmenu.sh is only ever run in a local session.

When connecting via SSH and launching zfsbootmenu from the shell, these
configuration variables are not accessible. The simple solution is to
write the configuration into a file that can be sourced by both
zfsbootmenu-countdown.sh and zfsbootmenu.sh to populate their
environments regardless of the way they are called.

The boot-environment list has been renamed from "env" to "bootenvs" to
avoid confusion with the "environment" file.