Use fzf's execute[] function to render draw_diff
Instead of exiting the fzf snapshot browser, bind a key combination to the execute[] function in fzf, allowing you to keep your place in the list after viewing a diff. zfs-chroot does not correctly function here, because it loses access to the backing TTY. Bash/dash will not write command output to the fzf alternate screen. Discovering the tty is easy for local consoles (control_term from -exec), but is much more difficult for connections over SSH.