Allow references to parent properties in org.zfsbootmenu:commandline
Any reference to "%{parent}" in org.zfsbootmenu:commandline will be
replaced with the value of the same property on the parent filesystem
(with parent references above recursively expanded), allowing easy
specification of common options at a mutual parent of two BEs and
overrides or additions of individual options per-BE. The value of
%{parent} is always an empty string on a root filesystem.

This is not intended to be sophisticated, and %{parent} appearing within
other words will be replaced regardless. The assumption is that
%{parent} is unique enough and will not conflict with real KCL options,
so dumb global replacement is sufficient.