Small quality of life improvements
When console=ttyS0,1152800n8 is passed in to the kernel it does not
directly translate to a /dev entry. Simply strip the ,value off of the
console=x,y parameter and set that as our /dev entry.

When a pool import fails for some reason, log the text from the zpool
import process at the debug log level.