Relax insmod spl failures, try to load zfs.ko
Instead of manually loading spl.ko every time, try to load it via
modprobe. Iff that fails, use modinfo to find the path to spl.ko and try
to load it with spl_hostid=0. Since udevd could have already loaded zfs
(and by extension, spl), insmod failures are not hard failures. Warn
that the module couldn't be loaded along with the error text and then
try to load zfs.ko anyway.

This moves all of the on-disk / /sys hostid manipulation to after the
modules have been loaded. This ensures that regardless of how spl was
loaded, /sys/module/spl/parameters/spl_hostid is set to 0 to force
/etc/hostid to take precedence.