Allow zfsbootmenu to takeover a running instance
Instead of requiring a user to rm /zfsbootmenu/active, running 'zfsbootmenu' or 'zbm' from the recovery shell will automatically stop an already-running copy of /bin/zfsbootmenu. It will also block the auto-boot countdown sequence that might be executing on the main tty (tty0, ttyS0, etc). 

If the running instance can't be killed for some reason (/zfsbootmenu/active still exists), zfsbootmenu will spin with the option to exit to a recovery shell until that file is removed.