Enable snapshot <> snapshot diffing
This change adds in the ability to diff a snapshot against another
snapshot, instead of always against the current state of the filesystem.

The tab key is used to select one or two snapshots that can be passed
into the diff browser.

If two snapshots are selected and a non-diff action is chosen, the
second snapshot (as determined by fzf) is discarded from the return
value.

Since there can potentially be confusion as to which snapshot was
selected, the clone/duplicate screens now show the snapshot to be
cloned/duplicated at the top of the screen.

The diff browser screen no longer shows the snapshot name as the input
prompt - a generic '> ' is instead used.

Lastly, zfsbootmenu-preview now accepts an optional third argument to be
shown on a 3rd preview line. This can be used to provide additional /
relevant help on the screen - such as noting the tab key can be used to
select two snapshots for diffing.