Replace mod+i with mod+j
ctrl+i is interpreted as the tab key. For multi-selection support on the
snapshot screen, the tab key needs to be used. If --expect=ctrl-i is
set, the tab key becomes a shortcut key to the chroot handler.

Switch to ctrl-j on the primary and snapshot screens and then update
documentation to reflect the change.