Deprecate KCL in /etc/default/{zfsbootmenu,grub}
ZBM will complain when reading kernel command-line parameters from
/etc/default/{zfsbootmenu,grub} within a BE rather than from the ZFS
property org.zfsbootmenu:commandline. By default, the contents of these
files will be migrated to a ZFS property on the BE after the warning is
shown. This will happen once for each BE that provides a configuration
file and no property. The file /etc/default/zfsbootmenu will be removed
from the BE if that was the source for the KCL; /etc/default/grub is
never removed.

By pressing Esc, the user can ignore the deprecation warning without
attempting to migrate the KCL configuration. This allows ZBM to operate
as before. When the warning is ignored, ZBM will export an environment
variable and write to /etc/zfsbootmenu.conf in the initramfs to suppress
subsequent warnings until the next reboot.

Co-authored-by: Andrew J. Hesford <ajh@sideband.org>
Co-authored-by: Zach Dykstra <dykstra.zachary@gmail.com>