Allow snapshot creation / move logic to function
Some times it's helpful to be able to create a snapshot from the
bootloader, outside of the OS. On the snapshot screen, there's now a
MOD+N option to create a (N)ew snapshot. It rides along the coat-tails
of most of the snapshot cloning/duplication logic, as that has checks
for everything we also care about.

The MOD+S handler in zfsbootmenu.sh has grown quite large, with the
creation of quite a few global variables. The bulk of this is now hidden
inside snapshot_dispatcher(), which can use local variables for all of
the space calculations.