Handle empty snapshots list with no-op item
fzf requires something to be selected for --expect keybinds to work. If
there are no snapshots, the fzf list is empty and MOD+N can't be
triggered.

Detect if there are no snapshots and feed 'No snapshots available' to
the snapshot pager. When handling the key + selection after
draw_snapshots() returns, ensure that the selected text has '@' in it,
otherwise ignore any action other than creating a new snapshot.