Try to inherit key (encryptionroot) in duplicate_snapshot
A raw send (necessary to preserve properties and avoid a decryption pass
on encrypted filesystems) will cause the target to be its own encryption
root, which is surprising and probably undesirable. If the source is not
its own encryption root, make the target inherit its key from a (common)
parent.