Use file-based lookup system for commandline values
Instead of parsing the kernel command line options every time
get_zbm_arg is called, parse them once and store any values in a file
based on the parameter name. get_zbm_arg now simply checks for the
argument file, and if found, echoes (returns) the value in the file.

If a value is found, get_zbm_arg returns 0, so trust this return code in
parse-commandline.sh.

Co-authored-by: Zach Dykstra <dykstra.zachary@gmail.com>
Co-authored-by: Andrew J. Hesford <ajh@sideband.org>